refactor: Extract components for taxonomy settings
