+++ src/FormTaxonomies/resources/form-builder/form-tags.tsx
@@ -3,6 +3,7 @@ import {debounce} from "@wordpress/compose";
 import apiFetch from '@wordpress/api-fetch';
 import {FormTokenField} from "@wordpress/components";
 import {getInitialFormTags} from "./windowData";
+import {useState} from "react";
 
 /**
  * @note Not sure why it won't let me use a state hook here.
